Separacao das bibliotecas, execucao de arquivos (./) 'arquivo para testes> sum.c'
